Remove watched directories
a3m.server.packages.Package now handles both Transfer and SIP objects
transparently, but still treated as different entities because the
underlying data model and client modules need both.

A practical way to see this is in the active queue where now we have one
active package across both transfer and ingest. There is no transition
between two packages anymore.

DirectoryClientScriptJob learns a new callback that ensures that Package is
signed to it starts behaving as a SIP. We may want to handle this differently
in the future but I needed a couple of things: come up with a solution that
does not involve removing the Transfer model (lots of work) and having the
ability to instantiate a Package that knows the UUID of its SIP so it can be
returned immediately to the user, hiding the Transfer which is less useful.