Start end-to-end integration
This commit adds a new module (`a3m.server.enduro`) that knows how to
perform as an activity worker using Cadence as the broker and Enduro as
the workflow scheduler.

It also include some improvements in the CLI client and the status poller.
It restores store_aip to upload the package to S3.